Specialcase UAP == windows (#7893)
Fixes #6150

Context
We verify that the target platform is appropriate for the target platform identifier, but that check is too specific; it doesn't permit equivalent strings from matching. So far, we only know of UAP being a windows app, hence valid for windows. This fixes that instance.

Changes Made
Testing
Tried to reproduce this with and without this change. The error changed, and I don't think it's related, but I'm keeping this a draft PR anyway because I'm not confident about this.