Fallback when server client fails due to mutex connection timeout error. (#8024)
Fixes #7993

Summary
In the MSBuild server, we identify the state of the server via mutexes. Sometimes, for reason yet unknown to us, mutex could throw the exception System.IO.IOException: Connection timed out. When this occurs, we fallback to old behavior building without server. We fixed some of those in #8000, but now found more situations when this happens.

Customer Impact
MSBuild non-Windows users could have intermittent error when building with dotnet build.
This does not affect Visual Studio.

Regression?
Yes, this is a regression.

Testing
Unit tests.

Risk
Low risk. The fix adds additional try-catch blocks to process this situation.

Code Reviewers
[TODO]

Description of the fix
Add a try-catch block to catch and process the IOException exception when mutexes are used.
Add a new client exit type for this kind of error.