Drop the net35 asset from the StringTools package
As discussed offline, the StringTools package is currently not authored correctly as the StringTools assembly is differently named between target frameworks. More precisely, the net35 assembly is named differently than the rest of the builds: Microsoft.NET.StringTools.net35.dll. Reason for that is that this assembly is binplaced into a common folder with another tfm build of the same library.

By dropping the net35 asset from the package, assembly FileNotFoundExceptions are avoided.