Improve SDK resolver error reporting if none of the SDK resolvers attached are able to resolve a particular SDK (#7882)
Permits SDK resolvers to return an enumerable of messages in addition to the warnings and errors. Messages are ignored if we end up finding an SDK resolver that can resolve the SDK, but they are aggregated and printed out all at once in an InvalidProjectFileException if we fail to resolve any SDK.

This will need changes to other SDK resolvers for it to be fully effective. If this is merged, I will attempt the requisite changes to the NuGetSDKResolver and maybe the C++ SDK resolver.

Note also that this includes a breaking change, but I suspect it's seldom used, perhaps not used at all by anyone outside Microsoft.

Co-authored-by: Forgind <Forgind@users.noreply.github.com>