extract common logging functionality into ApiSymbolExtensions (#29379)
This change refactors APICompat and GenAPI to use a common set of logging abstractions. These include `ILog`, `ISuppressableLog`, `MSBuildLog`, `ConsoleLog`, `SuppressableConsoleLog`, `SuppressableMSBuildLog`, and `SuppressableTestLog`.

```
ILog
├─ConsoleLog────────────┬─SuppressableConsoleLog
├─ISuppressableLog──────┤
│ └─SuppressableTestLog │
└─MSBuildLog────────────┴─SuppressableMSBuildLog
```

Because we don't want the console logger to depend on `Microsoft.Build.*` packages, we exclude files that depend on it from `ApiSymbolExtensions`'s build. The depending packages (in this case `GenAPI.Task` and `ApiCompat.Task`) are required to include those files explicitly.