Parse invalid property under  target (#8190)
Fixes #5773

Context
Put a property inside a target without an enclosing PropertyGroup and build. Error informing me that is an invalid child element 

Changes Made
Add one condition when parse the element under target. If the element has inner text and no other child elements except text, then this should be a property and throw invalid child element, the error is "MSB4067: The element &lt;{0}&gt; beneath element &lt;{1}&gt; is unrecognized. If you intended this to be a property, enclose it within a &lt;PropertyGroup&gt; element"

Tests
ReadInvalidPropertyUnderTarget