Stop dumping stack for BadImageFormatException (#8271)
Fixes #6224

Context
See issue -- a change was made to add more detail when RAR fails to resolve a reference due to encountering a BIFE. This added too much detail -- not only the message from the inner exception, but also the callstack. We should not dump callstack when there is no bug in MSBuild (or a task/logger).

Oddly enough I hit this again #8140 and didn't remember the original discussion, and therefore again assumed there was a bug.

Changes Made
Fix BIFE to include the message from any inner exception. (There was a suggestion in the issue that it also include the exception type, but I don't know of any case where the type is necessary in order to disambiguate what happened, and we don't have precedent for doing it. It would also be ugly.)
Revert the original change, now unnecesssary.

Testing
Added test that fails without the fix.