Disable MotW check on core in GenerateResource (#8234)
Fixes #7946

Context
The GenerateResource task previously would deserialize and reserialize resource files. It used BinaryFormatter, which is known to be unsafe; to mitigate that security risk, we added a check for the mark of the web on resource files; files with that mark are considered unsafe, and those without are considered safe to be deserialized. On core, we have a separate solution: don't deserialize and reserialize them.

More recently, in work to unify code paths, we effectively enabled the check on core. There remains a check that we're on windows, so it only affected windows computers, and most windows devices could accomplish that fine. Nanoservices, however, did not have the COM API that we called as part of the "IsDangerous" check. This led it to crash.

Customer Impact
Using a (windows nanoserver) container, customers cannot build projects that include certain resource files. They instead see a confusing error about a missing COM API.

Risk
Low

Testing
I verified that setting the registry key resolved the issue. Though that's a different change, it leads to the same code path.

Code Reviewers
rainersigwald

Description of the fix
I added an early return on core to get away from the error.

Notes
The Mark of the Web check was originally introduced as a security measure, but this code path did not initially have it and doesn't need it. This change passed a security review.

Thanks to @asalvo for finding the problematic change!