Avoid removing outer -> inner graph edge (#8267)
Originally when the graph code was written, a non-top-level multitargeting project would end up never actually dispatching to the inner builds, so that edge was removed from the graph. For example if A referenced B and B was multitargeting, A would call B(outer) with GetTargetFrameworks, and then call B(inner) with Build. B(outer) never called B(inner) (again, when B wasn't an entry point).

However, at this point GetTargetFrameworks on the outer build actually does call into the inner builds, specifically the GetTargetFrameworksWithPlatformForSingleTargetFramework target. See:

image

Beyond the correctness issue, the practical impact of this is that this causes graph builds to potentially attempt to schedule outer builds before inner builds, which become blocked on the GetTargetFrameworksWithPlatformForSingleTargetFramework target. This causes MSBuild to execute that target on-demand, which in turn leads to an odd behavior where the BuildManager attempts to schedule the inner build, but the configuration already exists and the CreateUniqueGlobalProperty code path is hit, which eventually causes a double-build for the inner build down the line.

Example of bad graph execution:

A(outer) starts executing
GetTargetFrameworksWithPlatformForSingleTargetFramework is requested on A(inner)
Not in cache, so A(inner) executes
Some time later, A(inner) is scheduled by the graph build
The configuration already exists from (3), and the request already has a ProjectInstance from the graph. This causes CreateUniqueGlobalProperty to be called in BuildManager.ResolveConfiguration, which makes the build of A(inner) have a unique dummy global property.
A(inner) executes
Some time later, another project C starts executing
C depends on A, so ends up trying to build A(inner)
usually A(inner) would come from the result cache. However, it does not since the results of A(inner) have the dummy global property so it's actually a different configuration
A(inner) builds a second time. Uh oh!