[FancyLogger] Hide messages for completed projects during build, but show them again in summary (#8284)
The FancyLogger currently shows before each project: a list of all its high-priority messages, warnings and errors. As well as the currently running target and task. However this information might not be very useful and instead verbose for finishder projects.
Hence, high priority messages shold only be shown for currently running projects, and then again in the build summary (displayed at the end of the build)