Support laxer isolation mode `-isolate:Message` (#8249)
Context
Support a laxer isolation mode -isolate:MessageUponIsolationViolation (or its short form -isolate:Message) for builds unable to adhere to the restrictive isolation mode -isolate:True, piggybacking on the existing skipStaticGraphIsolationConstraints flag. Isolation violations are logged as messages.

Changes
Allow a new value for the -isolate switch: MessageUponIsolationViolation (short form Message).
In this new mode, only the TargetResults for top-level targets are serialized into the output cache. This is to mitigate the chances of isolation-violating dependency targets relying on state potentially mutated by cached dependency targets. (E.g., the definition of a property.) As such, it is still possible for a dependent project to have a config entry C in its override caches and still need to obtain BuildResults from C in isolation violation, requiring placing them in the override result cache instead of the current cache to maintain the invariant of no overlap configs across the override and current caches.
Preserve the IsolateProjects getter/setter for API backwards compat.