Set an environment variable indicating that the binary logger is enabled. (#8395)
Fixes #8394

Context
This is mainly so that the static graph-based restore can detect that the user enabled the binary logger so it can automatically provide a binary log for restore which is run out-of-proc so it has to be another file.

Changes Made
This sets MSBUILDBINARYLOGGERENABLED=true when the binary logger is enabled. I originally tried to have a new property on the EngineServices but that doesn't work for out-of-proc nodes. To make that work, each out-of-proc node would need to be told if the binary logger is in use for that particular build which seems like a lot overhead compared to just setting an environment variable.

Testing
Built locally and ran the bootstrap MSBuild to verify the environment variable is set.