Fix .NET 3.5 InternableString.GetHashCode to match the full implementation (#8340)
Fixes part of 8329

Context
The 3.5 version of InternableString uses a simpler hash code calculation, the results of which were not matching the 4.72/Core version. This manifested as a failing unit test RetainsLastStringWithGivenHashCode in VS.

Changes Made
Made the routine return the same numbers as the other implementation. It fixed the UT and also made the hash code "better" as previously we were shifting instead of rotating bits.

Testing
Existing unit tests (previously failing).

Notes
It would be nice to figure out how to run 3.5 tests in CI.