dotnet/msbuild#6521 `ObjectModelHelpers` refactoring (#8327)
Fixes #6521

Context
Did suggested refactoring in the issue.
NullMetadataOnLegacyOutputItems_InlineTask won't be fixed as it tests Deprecated code, and change is needed there.

Changes Made
loggers are now passed to evaluation
ability to set logger verbosity to MockLogger and when using BuildProjectExpectSuccess and BuildProjectExpectFailure
same loggers are reused during build when using BuildProjectExpectSuccess and BuildProjectExpectFailure
registering logging is now done before the project is evaluated
improved doc and formatting in changed files
disposing of ProjectCollection in BuildProjectExpectSuccess and BuildProjectExpectFailure
Testing
Testing only changes

Notes
There are other calls to ObjectModelHelpers.CreateInMemoryProject which do not dispose ProjectCollection. Those needs to be fixed separately.