New Trace overloads without array allocation and boxing (#8345)
Fixes #7815

Context
CommunicationUtilities.Trace method uses String.Format with array of arguments. Allocation of this array happens even when the tracing is not enabled.

Changes Made
Added new overloads avoiding array allocation and boxing for max 3 arguments.

Testing
Traces are written to file when tracing is enabled.

Notes
It's difficult to precisely measure allocation savings in this case (expectations are low), but complexity of the added code code is low.