Enable CA1852: Seal internal types (#8389)
Context
Types that are not designed to be inherited from should be sealed. This is general defensive goodness, and also has perf implications because it enables some runtime/JIT optimizations. A new analyzer was recently added for this.

Changes Made
Enabled the analyzer and fixed the code with dotnet format.

Testing
Manual review of the changes.

Notes
dotnet format also rectified our namespace imports. I don't believe there's any controversy in that and the change is low risk so I left it in. For the actual sealing:

It missed one class because it was ifdefed out in Core. I fixed it manually.
It resulted in a handful of build errors where the compiler didn't like protected members in sealed types. I fixed those manually.
The analyzer doesn't understand InternalsVisibleTo, by design. This is not causing any issues for us now as we don't derive from internal types in other assemblies. Our InternalsVisibleTo all point to our own assemblies built in this repo so this is a non-breaking change (unless hacks, reflection, etc. but the bar is certainly not that high).
I am not including changes to files under src/Deprecated in this PR.