Turn off static graph restore. (#8498)
Our CI builds fails because of bug https://github.com/NuGet/Home/issues/12373. 
It is fixed in https://github.com/NuGet/NuGet.Client/pull/5010. 
We are waiting for it to flow to CI machines. Meanwhile this PR applies a workaround.

Note: This PR needs to be reverted once it happens.