Add explicit deployment tools dependency (#31129)
dotnet/sdk depends on deployment tools. Add dependencies on the specific package and, because the source build intermediate versioning does not match (Microsoft.Deployment* is different), the source-build intermediate.

This brings in a 7.0 dependency on System.Text.Json, otherwise downgrades will occur. Upgrade this repo to 7.0.0 across that an a few other dependencies. MSBuild has already moved to the 7.0.0 version so no issue is expected.

Remove an extra unneeded property for NewtonSoft.Json