Don't look in root for tools config
When use this command in Windows system, The process will search for tool information in the .config\dotnet-tools.json file from the bottom up and download it.

For Windows, the program does not check the root folder for the tool manifest if the Environment Variable `CHECK_MANIFEST_IN_ROOT" is not set.

For Linux/MacOS, the program checks the root folder for the tool manifest.