Tell bot to only post when a PR is opened or reopened and let it target more servicing branches (#8583)
This updates the bot that currently posts whenever any action is taken on a PR targeting 17.4 (i.e., opened, closed, reopened, commented on, etc.) to only post when the action is to open or reopen a PR. It also extends the branches for which this takes effect to include a number of vs* branches, including some that don't currently exist.

I think this is the final or penultimate iteration. Specifically, I don't think it will fire if you target main, then retarget to vs* without closing and reopening it...unfortunately, retargeting isn't something fabric bot understands. We can add "label added:Servicing-approved" to the list of options as a proxy. That isn't perfect, but it would make it slightly more robust and slightly noisier. I'm open to doing that or not doing that.