Prefer sln-defined platforms for command-line builds over dynamic platfor resolution (#8589)
Today dynamic platform resolution is inconsistent due to the condition being based on `$(BuildingInsideVisualStudio)`, which is obviously only set in VS. Sln-based command-line builds wouldn't have that set though, so dynamic platform resolution would end up running. The comment on `_GetProjectReferencePlatformProperties` implies that sln-provided platforms should be used instead though, so this change switches the condition to check `$(CurrentSolutionConfigurationContents)` instead to make the experience consistent when building a sln in VS or command-line.