Fix graph target list for multitargetting projects (#8565)
This change fixes graph builds for multitargetting projects using BuildProjectReferences=false

Previously, when using BuildProjectReferences=false the graph would consider the outer build calling GetTargetPath on the inner builds. However, it actually calls Build when dispatching to the inner builds, and those call GetTargetPath on their dependencies. In effect this is causing graph builds to be non-graph builds since the target list is not accurately described.

This change ensures the correct behavior by separating the ProjectReferenceTargets logic for single and multi-targetting projects.