Avoid overly-specific cast in bulk metadata copy (#8646)
The return value of `ITaskItem.CloneCustomMetadata` is an `IDictionary`,
which is generally (in modern MSBuild) backed by a
`Dictionary<string, string>`, but can be (when given an item from a
net35 taskhost) a `Hashtable`. In the latter situation, casting entries
to `KeyValuePair<,>` fails, because they conform only to
`DictionaryEntry`.

Use that less-well-typed approach--the casts were present in the pre-
bulk-edit version of the code. Fixes #8645.