NoProfile in ToolTaskThatTimeoutAndRetry test
This failed for me in an unrelated PR and on my dev box and
at least on my dev box it was because sometimes my PowerShell
profile took longer than 1 second to process.

Pass -NoProfile to avoid that time entirely when it's not necessary.

In addition, save the engine's log to the test log to give some clues
if the test fails again.