GetReferenceAssemblyPaths continues on error in design-time builds
This fixes an issue where VS design-time builds would fail when reference assemblies could not be found.

By allowing the design-time build to continue, the .NET Project System will the nominate a restore which may bring in a package that provides those reference assemblies.

Without this addition, the task will fail and the build will end early, such that the restore does not occur and progress is not made.

This helps when users do not have targeting packs installed (such as for out-of-support versions of .NET Framework, like v4.5). With this change, a reference assembly package (like `Microsoft.NETFramework.ReferenceAssemblies.net45`) may be downloaded for the user to compile against.