Unify "used response file" messages
In #8608 I introduced a new string for this, but it's redundant with the one we already had that we only used to log to the console.

Dropping the "log directly to console" approach because we now (18 years later) have deferred messages, the engine API that someone didn't want to add.