Prevent restore loop in VS
The `GetReferenceAssemblyPaths` target adds a `PackageReference` to `Microsoft.NETFramework.ReferenceAssemblies` when needed.

In VS, this new package causes VS to nominate a NuGet restore then perform another design-time build.

That second design-time build was previously determining that the `PackageReference` was no longer needed, then removing it. The task's `FullFrameworkReferenceAssemblyPaths` output parameter contains the full path of the reference assembly folder in such a case, which fails the empty-string test.

This change tests whether the output parameter is a reference to the package, and if so allows the `PackageReference` to remain.