Add PackageReference to new Roslyn package (#29787)
If the version of the SDK is different from what is used in VS, it can cause numerous problems as different components expect different versions hence APIs and functionality. This resolves the problem by providing a way for the SDK to acquire a .NET Framework version of Roslyn with the same version as that in the SDK. (The version is specified in the dotnet/installer repo.)

That functionality is turned off for non-Windows machines, as they do not support .NET Framework. It also provides a warning if you try to reference the package directly, as that is not intended behavior.