Close API gap where ProjectInstance ignores ProjectOptions.DirectoryCacheFactory (#8746)
Context
When ProjectOptions.DirectoryCacheFactory was introduced only Project support was implemented. ProjectInstance ignores the field and while currently not a required functionality, it is a minor API gap.

Changes Made
Made ProjectInstance use DirectoryCacheFactory if passed in ProjectOptions.

Removed Project.GetDirectoryCacheForEvaluation and instead plumbed the factory all the way down to Evaluator.
Tweaked the unit test project to use Condition on Import rather than on ItemGroup to not expose differences between Project and ProjectInstance in the treatment of conditional items.
Testing
Existing updated unit test.